client/requirements.txt
flwr==1.7.0
flwr_datasets
numpy
pyyaml
